Improve sync file creation error handling

+++ urbackupserver/IncrFileBackup.cpp
@@ -1524,9 +1524,11 @@ bool IncrFileBackup::doFileBackup()
 			else
 			{
 				ServerLogger::Log(logid, "Error creating sync file at " + backuppath_hashes + os_file_sep() + sync_fn+". Not setting backup to done.", LL_ERROR);
+				c_has_error = true;
 			}
 
-			if (ServerCleanupThread::isClientlistDeletionAllowed())
+			if (!c_has_error
+				&& ServerCleanupThread::isClientlistDeletionAllowed())
 			{
 				Server->deleteFile(clientlist_name);
 			}
@@ -1614,9 +1616,11 @@ bool IncrFileBackup::doFileBackup()
 		else
 		{
 			ServerLogger::Log(logid, "Error creating sync file at " + backuppath_hashes + os_file_sep() + sync_fn+". Not setting backup to done", LL_ERROR);
+			c_has_error = true;
 		}
 
-		if (ServerCleanupThread::isClientlistDeletionAllowed())
+		if (!c_has_error
+			&& ServerCleanupThread::isClientlistDeletionAllowed())
 		{
 			Server->deleteFile(clientlist_name);
 		}
